Recoded SMT automaton transformation
The old transformation wasn't capable of handling nodes which were
used inside a location. This is essential since otherwise we would
have to pull out nodes and rewrite them so that they only advance
their state if the corresponding location is not active.

Now we do something similar but on the level of the transformation.
Every statement can now have an activation condition. If it holds,
the normal data flow is executed. Otherwise local variables get
some (more or less) undefined values and states are just transported
to the next cycle. The same holds for active locations.
The undefined values do no harm, since an activation condition always
holds for a in a whole. So the local variables are never accessed.

The transformation now works by directly translating an automaton
into ite chains instead of translating first into enums.