Corrected translation of Scade automata
We also now build up the graph first and then transform
it.

The translation of weak transitions was not correct.
The interaction with strong transitions is a bit trickier.
If a state s is entered in a cycle by a strong transition
and a condition on a weak transition leading out of s
is true, that transition should not be activated. This
means that we have extend the condition for a weak
transition by conditions that hold only if s hasn't been
entered by a strong transition in the current cycle.